Move core-navigation to navigation:navigation-activity
To make room for more navigation-related modules
