--- core-navigation/build.gradle.kts
@@ -1,8 +1,0 @@
-plugins {
-  id("hedvig.android.library")
-  id("hedvig.android.ktlint")
-}
-
-android {
-  namespace = "com.hedvig.android.navigation"
-}
