+++ navigation/navigation-activity/src/main/AndroidManifest.xml
