+++ settings.gradle.kts
@@ -45,7 +45,6 @@ include(":core-common-test")
 include(":core-datastore")
 include(":core-datastore-test")
 include(":core-design-system")
-include(":core-navigation")
 include(":core-resources")
 include(":core-ui")
 include(":datadog")
@@ -62,5 +61,7 @@ include(":hanalytics:hanalytics-feature-flags-test")
 include(":hedvig-language")
 include(":hedvig-market")
 include(":micro-apps:design-showcase")
+include(":navigation:navigation-activity")
+include(":navigation:navigation-compose-typed")
 include(":notification-badge-data")
 include(":testdata")
